using floats with 10 decimal points

+++ week_01/src/floats.c
@@ -6,11 +6,11 @@
 int main()
 {
     // Promt user for x
-    int x = get_float("x: ");
+    float x = get_float("x: ");
 
     // Promt user for y
-    int y = get_float("y: ");
+    float y = get_float("y: ");
 
     // Perform arithmetic
-    printf("%f divided by %f is %i\n", x, y, x / y);
+    printf("%f divided by %f is %.10f\n", x, y, x / y);
 }
